Fixed error with login and ui
There was an error, where the header and the UserButton showed the wrong links while not logged in. Now, when the user logges in, <SESSION.loggedin> is set to true and to false if the user logges out. <SESSION.loggedinUserId> holds the id from the logged in user and if there is no user logged in -1